requirements.txt
# Data manipulation
pandas==2.1.4
numpy==1.26.4
pyarrow==14.0.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# API requests
requests==2.31.0
python-dotenv==1.0.0

# Machine Learning (compatible Python 3.12)
scikit-learn==1.4.0
tensorflow==2.19.1
keras==3.7.0

# Azure (para después)
azure-ai-ml==1.12.0
azure-identity==1.15.0

# Jupyter
jupyterlab==4.0.9
ipykernel==6.27.1

requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.2
pyarrow==16.1.0
py7zr==0.21.0
tqdm==4.66.4
